Giordano's contributions: What do you know? I did stuff
generalized the initial electron placement function;
created algorithm to move electrons without overlapping in each time
step;
found pythtb to be explored though it may largely just serve to display
the dispersion relation;

Work to be done is listed below:

Still need to create a test case but I think a good one would be the
fraction of jumps over timesteps compared to probabilities of jumping
and how this changes with the initial distribution of electrons.
could make the number of electrons dependent upon temperature.
need to confirm the dependence of hopping probabilities with temperature
and change the motion function to take the probabilities as arguements.

Could be fancy and incorporate a sliding bar to change parameters